+++ modules/src/ics04_channel/msgs/chan_open_ack.rs
@@ -19,12 +19,12 @@ const TYPE_MSG_CHANNEL_OPEN_ACK: &str = "channel_open_ack";
 ///
 #[derive(Clone, Debug, PartialEq)]
 pub struct MsgChannelOpenAck {
-    port_id: PortId,
-    channel_id: ChannelId,
-    counterparty_channel_id: ChannelId,
-    counterparty_version: String,
-    proofs: Proofs,
-    signer: AccountId,
+    pub port_id: PortId,
+    pub channel_id: ChannelId,
+    pub counterparty_channel_id: ChannelId,
+    pub counterparty_version: String,
+    pub proofs: Proofs,
+    pub signer: AccountId,
 }
 
 impl MsgChannelOpenAck {
@@ -70,6 +70,9 @@ impl Msg for MsgChannelOpenAck {
         // Nothing to validate
         // All the validation is performed on creation
         Ok(())
+    }
+    fn type_url(&self) -> String {
+        "/ibc.core.channel.v1.MsgChannelOpenAck".to_string()
     }
 
     fn get_signers(&self) -> Vec<AccountId> {