+++ modules/src/ics04_channel/msgs/chan_open_confirm.rs
@@ -19,10 +19,10 @@ const TYPE_MSG_CHANNEL_OPEN_CONFIRM: &str = "channel_open_confirm";
 ///
 #[derive(Clone, Debug, PartialEq)]
 pub struct MsgChannelOpenConfirm {
-    port_id: PortId,
-    channel_id: ChannelId,
-    proofs: Proofs,
-    signer: AccountId,
+    pub port_id: PortId,
+    pub channel_id: ChannelId,
+    pub proofs: Proofs,
+    pub signer: AccountId,
 }
 
 impl MsgChannelOpenConfirm {
@@ -64,6 +64,10 @@ impl Msg for MsgChannelOpenConfirm {
         // Nothing to validate
         // All the validation is performed on creation
         Ok(())
+    }
+
+    fn type_url(&self) -> String {
+        "/ibc.core.channel.v1.MsgChannelOpenConfirm".to_string()
     }
 
     fn get_signers(&self) -> Vec<AccountId> {