+++ modules/src/mock_client/error.rs
@@ -1,4 +1,0 @@
-use thiserror::Error;
-
-#[derive(Debug, Error)]
-pub enum MockError {}